refactor : vi, guide, user 엔티티 user 패키지로 통합
