+++ run/src/main/java/com/guide/run/global/security/user/CustomUserDetails.java
@@ -1,6 +1,6 @@
 package com.guide.run.global.security.user;
 
-import com.guide.run.user.entity.User;
+import com.guide.run.user.entity.user.User;
 import lombok.AllArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
