+++ run/src/main/java/com/guide/run/global/security/user/CustomUserDetailsService.java
@@ -1,7 +1,7 @@
 package com.guide.run.global.security.user;
 
 import com.guide.run.global.exception.user.resource.NotExistUserException;
-import com.guide.run.user.entity.User;
+import com.guide.run.user.entity.user.User;
 import com.guide.run.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
