+++ run/src/main/java/com/guide/run/user/dto/PersonalInfoDto.java
@@ -1,6 +1,6 @@
 package com.guide.run.user.dto;
 
-import com.guide.run.user.entity.User;
+import com.guide.run.user.entity.user.User;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
