+++ run/src/main/java/com/guide/run/user/entity/partner/PartnerId.java
@@ -1,4 +1,4 @@
-package com.guide.run.user.entity;
+package com.guide.run.user.entity.partner;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
