+++ run/src/main/java/com/guide/run/user/entity/user/User.java
@@ -1,4 +1,4 @@
-package com.guide.run.user.entity;
+package com.guide.run.user.entity.user;
 
 
 import com.guide.run.global.entity.BaseEntity;
