+++ run/src/main/java/com/guide/run/user/repository/PartnerRepository.java
@@ -1,12 +1,8 @@
 package com.guide.run.user.repository;
 
-import com.guide.run.user.entity.Partner;
-import com.guide.run.user.entity.PartnerId;
+import com.guide.run.user.entity.partner.Partner;
+import com.guide.run.user.entity.partner.PartnerId;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
-import java.util.Optional;
 
 public interface PartnerRepository extends JpaRepository<Partner, PartnerId> {
