+++ run/src/main/java/com/guide/run/user/repository/ViRepository.java
@@ -1,6 +1,6 @@
 package com.guide.run.user.repository;
 
-import com.guide.run.user.entity.Vi;
+import com.guide.run.user.entity.user.Vi;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ViRepository extends JpaRepository<Vi, String> {
