+++ run/src/main/java/com/guide/run/user/service/MypageService.java
@@ -2,7 +2,7 @@ package com.guide.run.user.service;
 
 import com.guide.run.global.exception.user.resource.NotExistUserException;
 import com.guide.run.user.dto.GlobalUserInfoDto;
-import com.guide.run.user.entity.User;
+import com.guide.run.user.entity.user.User;
 import com.guide.run.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
