+++ run/src/test/java/com/guide/run/user/service/UserServiceTest.java
@@ -1,9 +1,8 @@
 package com.guide.run.user.service;
 
 import com.guide.run.user.entity.type.Role;
-import com.guide.run.user.entity.User;
+import com.guide.run.user.entity.user.User;
 
-import com.guide.run.user.entity.Vi;
 import com.guide.run.user.repository.UserRepository;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -11,8 +10,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-
-import java.util.UUID;
 
 @SpringBootTest
 class UserServiceTest {
